Extract BillingContext from conversations.rs
